completed the following: extracted UpdateUserRequest from UserResponseDto to a separated class corrected update booking request file location
