+++ src/main/java/com/devbp/syncspace/controllers/UserController.java
@@ -1,6 +1,7 @@
 package com.devbp.syncspace.controllers;
 
 import com.devbp.syncspace.domain.dtos.CreateUserRequest;
+import com.devbp.syncspace.domain.dtos.UpdateUserRequest;
 import com.devbp.syncspace.domain.dtos.UpdateUserRequestDto;
 import com.devbp.syncspace.domain.dtos.UserResponseDto;
 import com.devbp.syncspace.domain.mappers.UserMapper;
@@ -61,7 +62,7 @@ public class UserController {
     @PutMapping(path = "/{id}")
     public ResponseEntity<UserResponseDto> updateUser(@PathVariable long id, @Valid @RequestBody UpdateUserRequestDto updateUserRequestDto) {
 
-        UserResponseDto.UpdateUserRequest updateUserRequest = userMapper.toUpdateUserRequest(updateUserRequestDto);
+        UpdateUserRequest updateUserRequest = userMapper.toUpdateUserRequest(updateUserRequestDto);
 
         UserResponseDto userResponseDto = userMapper.toDto(userService.updateUser(id, updateUserRequest));
