+++ src/main/java/com/devbp/syncspace/domain/dtos/UpdateBookingRequest.java
@@ -1,0 +1,4 @@
+package com.devbp.syncspace.domain.dtos;
+
+public class UpdateBookingRequest {
+}
