+++ src/main/java/com/devbp/syncspace/domain/entities/InvoiceItems.java
@@ -25,7 +25,7 @@ public class InvoiceItems {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "booking_id", nullable = false)
-    private Bookings bookings;
+    private Booking booking;
 
     @NotNull
     @Column(nullable = false)
@@ -46,11 +46,11 @@ public class InvoiceItems {
     public boolean equals(Object o) {
         if (o == null || getClass() != o.getClass()) return false;
         InvoiceItems that = (InvoiceItems) o;
-        return Objects.equals(invoice, that.invoice) && Objects.equals(bookings, that.bookings);
+        return Objects.equals(invoice, that.invoice) && Objects.equals(booking, that.booking);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(invoice, bookings);
+        return Objects.hash(invoice, booking);
     }
 }
