+++ src/main/java/com/devbp/syncspace/domain/mappers/UserMapper.java
@@ -1,5 +1,6 @@
 package com.devbp.syncspace.domain.mappers;
 
+import com.devbp.syncspace.domain.dtos.UpdateUserRequest;
 import com.devbp.syncspace.domain.dtos.UpdateUserRequestDto;
 import com.devbp.syncspace.domain.dtos.UserResponseDto;
 import com.devbp.syncspace.domain.entities.User;
@@ -15,6 +16,6 @@ public interface UserMapper {
     @Mapping(target = "updatedAt", source = "updatedAt")
     UserResponseDto toDto(User user);
 
-    UserResponseDto.UpdateUserRequest toUpdateUserRequest(UpdateUserRequestDto updateUserRequestDto);
+    UpdateUserRequest toUpdateUserRequest(UpdateUserRequestDto updateUserRequestDto);
 
 }
