+++ src/main/java/com/devbp/syncspace/services/BookingService.java
@@ -2,7 +2,7 @@ package com.devbp.syncspace.services;
 
 import com.devbp.syncspace.domain.BookingStatus;
 import com.devbp.syncspace.domain.dtos.CreateBookingRequest;
-import com.devbp.syncspace.domain.dtos.UserResponseDto;
+import com.devbp.syncspace.domain.dtos.UpdateBookingRequest;
 import com.devbp.syncspace.domain.entities.Booking;
 
 public interface BookingService {
@@ -13,7 +13,7 @@ public interface BookingService {
 
     Booking createBooking( CreateBookingRequest createBookingRequest);
 
-    Booking updateBooking(long id, UserResponseDto.UpdtaeBookingRequest updtaeBookingRequest);
+    Booking updateBooking(long id, UpdateBookingRequest updateBookingRequest);
 
     Booking confirmBookingStatus(long id, BookingStatus bookingStatus);
